Add PeerMessageSenderApi
